Separate "force fulfill" operation from the CommitInstantiated
While the RequestInstantiate fulfills `entry.[[Instantiate]]` promise
with the registry entry (as the returned value of the promise),
CommitInstantiated step 4 will fulfill it with the module instance proactively.

In this patch, we separate the "force fulfill" operation from the CommitInstantiated.
Move this operation to the FulFillInstantiate. And here, we will fulfill the `entry.[[Instantiate]]`
promise with the registry entry because the hander side requires it.